Aggiunta missioni, sistemazioni grafiche, riprogettazione del workflow
Suddivisione in missioni effettuata (gestite come cartelle all'interno della cartella foto), possibilità di salvare scontrini all'interno della missione e di visualizzarle a schermo intero.
Le activity sono state riprogettate e ne sono state aggiunte di nuove, ora il Main è la sezione dedicata alle missioni.
TODO: Al click del + nella sezione missioni creare la possibilità di aggiungerne una
