--- app/src/main/java/com/example/nicoladalmaso/gruppo1/DatabaseStrings.java
@@ -1,16 +1,0 @@
-package com.example.nicoladalmaso.gruppo1;
-
-/**
- * Created by nicoladalmaso on 28/10/17.
- */
-
-public class DatabaseStrings
-{
-    public static final String FIELD_ID="_id";
-    public static final String FIELD_NAME="nome";
-    public static final String FIELD_DESC="descrizione";
-    public static final String FIELD_IMAGE="immagine";  //PER SALVARE DIRETTAMENTE L'IMMAGINE
-    public static final String FIELD_DIRECTORY="percorso";
-    public static final String FIELD_DATE="data";
-    public static final String TBL_NAME="Scontrini";
-}
