+++ app/src/main/res/layout/activity_main.xml
@@ -1,57 +1,35 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.nicoladalmaso.gruppo1.MainActivity">
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="56dp"
-        android:layout_height="56dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginBottom="16dp"
-        android:layout_marginRight="16dp"
-        android:background="@color/colorPrimary"
-        android:clickable="true"
-        app:srcCompat="@mipmap/ic_add_white_24dp" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab1"
-        android:layout_width="56dp"
-        android:layout_height="56dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginBottom="80dp"
-        android:layout_marginRight="16dp"
-        android:clickable="true"
-        android:visibility="invisible"
-        app:backgroundTint="#D81B60"
-        app:srcCompat="@drawable/camera" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab2"
-        android:layout_width="56dp"
-        android:layout_height="56dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginBottom="140dp"
-        android:layout_marginRight="16dp"
-        android:clickable="true"
-        android:visibility="invisible"
-        app:backgroundTint="#8E24AA"
-        app:srcCompat="@mipmap/ic_image_white_24dp" />
-
-    <ListView
-        android:id="@+id/list1"
+    tools:context="com.example.nicoladalmaso.gruppo1.Missions"
+    android:background="#DFDFDF">
+    <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:listSelector="@android:color/transparent"
-        android:layout_alignParentTop="true"
-        android:divider="#00000000"
-        android:layout_centerHorizontal="true" />
+        android:layout_height="match_parent">
 
+        <ListView
+            android:id="@+id/listMission"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_alignParentEnd="true"
+            android:layout_alignParentTop="true"
+            android:divider="@null"
+            android:dividerHeight="0dp"/>
 
-</RelativeLayout>
+        <android.support.design.widget.FloatingActionButton
+            android:id="@+id/fab"
+            android:layout_width="56dp"
+            android:layout_height="56dp"
+            android:layout_alignParentBottom="true"
+            android:layout_alignParentRight="true"
+            android:layout_marginBottom="16dp"
+            android:layout_marginRight="16dp"
+            android:background="@color/colorPrimary"
+            android:clickable="true"
+            app:srcCompat="@mipmap/ic_add_white_24dp" />
+
+    </RelativeLayout>
+</LinearLayout>
