+++ app/src/main/res/layout/cardview.xml
@@ -6,7 +6,11 @@
     <android.support.v7.widget.CardView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="16dp">
+        android:layout_marginTop="10dp"
+        android:layout_marginLeft="8dp"
+        android:layout_marginRight="8dp"
+        android:layout_marginBottom="0dp"
+        android:elevation="8dp">
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
