+++ app/src/main/res/values/colors.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#E53935</color>
-    <color name="colorPrimaryDark">#D32F2F</color>
-    <color name="colorAccent">#F44336</color>
+    <color name="colorPrimary">#FF0056</color>
+    <color name="colorPrimaryDark">#D60052</color>
+    <color name="colorAccent">#FF0056</color>
     <color name="colorFAB1">#2196f3</color>
     <color name="colorFAB2">#E040FB</color>
